# Improve performance, fix bug in authorization schema