--- a/examples/example.js
+++ b/examples/example.js
@@ -20,7 +20,9 @@ fastify.listen({ port: 8000 }, (err) => {
 
 // Missing Header
 // autocannon http://127.0.0.1:8000/foo
-// Invalid Header
+// Invalid Bearer Type
+// autocannon -H authorization='Beaver key' http://127.0.0.1:8000/foo
+// Invalid Key
 // autocannon -H authorization='Bearer invalid' http://127.0.0.1:8000/foo
 // Valid Request
 // autocannon -H authorization='Bearer key' http://127.0.0.1:8000/foo