--- a/lib/verify-bearer-auth-factory.js
+++ b/lib/verify-bearer-auth-factory.js
@@ -48,7 +48,7 @@ module.exports = function verifyBearerAuthFactory (options) {
     }
 
     const key = header.substring(bearerType.length).trim()
-    let retVal
+    let retVal = false
     // check if auth function is defined
     if (auth && auth instanceof Function) {
       try {
@@ -62,17 +62,16 @@ module.exports = function verifyBearerAuthFactory (options) {
       retVal = authenticate(keys, key)
     }
 
-    const invalidKeyError = Error('invalid authorization header')
-
     // retVal contains the result of the auth function if defined or the
     // result of the key comparison.
     // retVal is enclosed in a Promise.resolve to allow auth to be a normal
-    // function or an async funtion. If it returns a non-promise value it
+    // function or an async function. If it returns a non-promise value it
     // will be converted to a resolving promise. If it returns a promise it
     // will be resolved.
     Promise.resolve(retVal).then((val) => {
       // if val is not truthy return 401
       if (val === false) {
+        const invalidKeyError = Error('invalid authorization header')
         if (verifyErrorLogLevel) request.log[verifyErrorLogLevel]('unauthorized: %s', invalidKeyError.message)
         if (contentType) reply.header('content-type', contentType)
         reply.code(401)