--- a/lib/verify-bearer-auth-factory.js
+++ b/lib/verify-bearer-auth-factory.js
@@ -37,17 +37,17 @@ module.exports = function verifyBearerAuthFactory (options) {
       reply.send(errorResponse(noHeaderError))
     }
 
-    const header = request.raw.headers.authorization
-    if (!header) {
+    const authorizationHeader = request.raw.headers.authorization
+    if (!authorizationHeader) {
       return authorizationHeaderErrorFn('missing authorization header')
     }
 
-    const type = header.substring(0, bearerType.length)
+    const type = authorizationHeader.substring(0, bearerType.length)
     if (type !== bearerType) {
       return authorizationHeaderErrorFn('invalid authorization header')
     }
 
-    const key = header.substring(bearerType.length).trim()
+    const key = authorizationHeader.substring(bearerType.length).trim()
     let retVal = false
     // check if auth function is defined
     if (auth && auth instanceof Function) {