--- a/lib/verify-bearer-auth-factory.js
+++ b/lib/verify-bearer-auth-factory.js
@@ -24,6 +24,9 @@ module.exports = function verifyBearerAuthFactory (options) {
     keys[i] = Buffer.from(keys[i])
   }
 
+  const bearerTypePrefix = bearerType + ' '
+  const bearerTypePrefixLength = bearerType.length + 1
+
   return function verifyBearerAuth (request, reply, done) {
     function authorizationHeaderErrorFn (errorMessage) {
       const noHeaderError = Error(errorMessage)
@@ -42,12 +45,12 @@ module.exports = function verifyBearerAuthFactory (options) {
       return authorizationHeaderErrorFn('missing authorization header')
     }
 
-    const type = authorizationHeader.substring(0, bearerType.length)
-    if (type !== bearerType) {
+    const type = authorizationHeader.substring(0, bearerTypePrefixLength)
+    if (type !== bearerTypePrefix) {
       return authorizationHeaderErrorFn('invalid authorization header')
     }
 
-    const key = authorizationHeader.substring(bearerType.length).trim()
+    const key = authorizationHeader.substring(bearerTypePrefixLength).trim()
     let retVal = false
     // check if auth function is defined
     if (auth && auth instanceof Function) {