--- a/lib/verify-bearer-auth-factory.js
+++ b/lib/verify-bearer-auth-factory.js
@@ -45,8 +45,7 @@ module.exports = function verifyBearerAuthFactory (options) {
       return authorizationHeaderErrorFn('missing authorization header')
     }
 
-    const type = authorizationHeader.substring(0, bearerTypePrefixLength)
-    if (type !== bearerTypePrefix) {
+    if (authorizationHeader.substring(0, bearerTypePrefixLength) !== bearerTypePrefix) {
       return authorizationHeaderErrorFn('invalid authorization header')
     }