--- a/lib/verify-bearer-auth-factory.js
+++ b/lib/verify-bearer-auth-factory.js
@@ -28,7 +28,7 @@ module.exports = function verifyBearerAuthFactory (options) {
   const bearerTypePrefixLength = bearerType.length + 1
 
   return function verifyBearerAuth (request, reply, done) {
-    function authorizationHeaderErrorFn (errorMessage) {
+    function handleUnauthorized (errorMessage) {
       const noHeaderError = Error(errorMessage)
       if (verifyErrorLogLevel) request.log[verifyErrorLogLevel]('unauthorized: %s', noHeaderError.message)
       if (contentType) reply.header('content-type', contentType)
@@ -42,11 +42,11 @@ module.exports = function verifyBearerAuthFactory (options) {
 
     const authorizationHeader = request.raw.headers.authorization
     if (!authorizationHeader) {
-      return authorizationHeaderErrorFn('missing authorization header')
+      return handleUnauthorized('missing authorization header')
     }
 
     if (authorizationHeader.substring(0, bearerTypePrefixLength) !== bearerTypePrefix) {
-      return authorizationHeaderErrorFn('invalid authorization header')
+      return handleUnauthorized('invalid authorization header')
     }
 
     const key = authorizationHeader.substring(bearerTypePrefixLength).trim()
@@ -73,12 +73,7 @@ module.exports = function verifyBearerAuthFactory (options) {
     Promise.resolve(retVal).then((val) => {
       // if val is not truthy return 401
       if (val === false) {
-        const invalidKeyError = Error('invalid authorization header')
-        if (verifyErrorLogLevel) request.log[verifyErrorLogLevel]('unauthorized: %s', invalidKeyError.message)
-        if (contentType) reply.header('content-type', contentType)
-        reply.code(401)
-        if (!addHook) return done(invalidKeyError)
-        reply.send(errorResponse(invalidKeyError))
+        handleUnauthorized('invalid authorization header')
         return
       }
       if (val === true) {