--- a/lib/verify-bearer-auth-factory.js
+++ b/lib/verify-bearer-auth-factory.js
@@ -27,26 +27,26 @@ module.exports = function verifyBearerAuthFactory (options) {
   const bearerTypePrefix = bearerType + ' '
   const bearerTypePrefixLength = bearerType.length + 1
 
-  return function verifyBearerAuth (request, reply, done) {
-    function handleUnauthorized (errorMessage) {
-      const noHeaderError = Error(errorMessage)
-      if (verifyErrorLogLevel) request.log[verifyErrorLogLevel]('unauthorized: %s', noHeaderError.message)
-      if (contentType) reply.header('content-type', contentType)
-      reply.code(401)
-      if (!addHook) {
-        done(noHeaderError)
-        return
-      }
-      reply.send(errorResponse(noHeaderError))
+  function handleUnauthorized (request, reply, done, message) {
+    const noHeaderError = Error(message)
+    if (verifyErrorLogLevel) request.log[verifyErrorLogLevel]('unauthorized: %s', noHeaderError.message)
+    if (contentType) reply.header('content-type', contentType)
+    reply.code(401)
+    if (!addHook) {
+      done(noHeaderError)
+      return
     }
+    reply.send(errorResponse(noHeaderError))
+  }
 
+  return function verifyBearerAuth (request, reply, done) {
     const authorizationHeader = request.raw.headers.authorization
     if (!authorizationHeader) {
-      return handleUnauthorized('missing authorization header')
+      return handleUnauthorized(request, reply, done, 'missing authorization header')
     }
 
     if (authorizationHeader.substring(0, bearerTypePrefixLength) !== bearerTypePrefix) {
-      return handleUnauthorized('invalid authorization header')
+      return handleUnauthorized(request, reply, done, 'invalid authorization header')
     }
 
     const key = authorizationHeader.substring(bearerTypePrefixLength).trim()
@@ -73,7 +73,7 @@ module.exports = function verifyBearerAuthFactory (options) {
     Promise.resolve(retVal).then((val) => {
       // if val is not truthy return 401
       if (val === false) {
-        handleUnauthorized('invalid authorization header')
+        handleUnauthorized(request, reply, done, 'invalid authorization header')
         return
       }
       if (val === true) {