--- a/test/verify-bearer-auth-factory.test.js
+++ b/test/verify-bearer-auth-factory.test.js
@@ -218,6 +218,29 @@ test('hook accepts correct header and alternate Bearer', (t) => {
   })
 })
 
+test('hook throws if header misses at least one space after bearerType', (t) => {
+  t.plan(2)
+
+  const request = {
+    log: { error: noop },
+    raw: {
+      headers: { authorization: `Bearer${key}` }
+    }
+  }
+  const response = {
+    code: () => response,
+    send
+  }
+
+  function send (body) {
+    t.ok(body.error)
+    t.match(body.error, /invalid authorization header/)
+  }
+
+  const hook = verifyBearerAuthFactory(keys)
+  hook(request, response)
+})
+
 test('hook accepts correct header with extra padding', (t) => {
   t.plan(1)