… test/decorate-with-logger.test.js