While programming I realized we have a bug in the code. We did not check if between `Bearer`and key was atleast one space. This bug is fixed, see unit tests.

before:
Missing Header:
```sh
autocannon http://127.0.0.1:8000/foo
0 2xx responses, 271167 non 2xx responses
271k requests in 10.01s, 59.9 MB read
```

Invalid BearerType:
```sh
autocannon -H authorization='Beaver key' http://127.0.0.1:8000/foo
0 2xx responses, 295334 non 2xx responses
295k requests in 11.01s, 65.3 MB read
```

Invalid key:
```sh
autocannon -H authorization='Bearer invalid' http://127.0.0.1:8000/foo
0 2xx responses, 274488 non 2xx responses
274k requests in 11.01s, 60.7 MB read
```

Valid Request:
```sh
autocannon -H authorization='Bearer key' http://127.0.0.1:8000/foo
272k requests in 11.01s, 52.6 MB read
```

after:
Missing Header:
```sh
autocannon http://127.0.0.1:8000/foo
0 2xx responses, 300072 non 2xx responses
300k requests in 11.01s, 66.3 MB read
```

Invalid BearerType:
```sh
autocannon -H authorization='Beaver key' http://127.0.0.1:8000/foo
0 2xx responses, 291809 non 2xx responses
292k requests in 11.01s, 64.5 MB read
```

Invalid key:
```sh
autocannon -H authorization='Bearer invalid' http://127.0.0.1:8000/foo
0 2xx responses, 326559 non 2xx responses
327k requests in 11.01s, 72.2 MB read
```

Valid Request:
```sh
autocannon -H authorization='Bearer key' http://127.0.0.1:8000/foo
404k requests in 11.01s, 77.9 MB read
```


<!--
Thank you for your pull request. Please provide a description above and review
the requirements below.

Bug fixes and new features should include tests and possibly benchmarks.

Contributors guide: https://github.com/fastify/fastify/blob/master/CONTRIBUTING.md

By making a contribution to this project, I certify that:

* (a) The contribution was created in whole or in part by me and I
  have the right to submit it under the open source license
  indicated in the file; or

* (b) The contribution is based upon previous work that, to the best
  of my knowledge, is covered under an appropriate open source
  license and I have the right under that license to submit that
  work with modifications, whether created in whole or in part
  by me, under the same open source license (unless I am
  permitted to submit under a different license), as indicated
  in the file; or

* (c) The contribution was provided directly to me by some other
  person who certified (a), (b) or (c) and I have not modified
  it.

* (d) I understand and agree that this project and the contribution
  are public and that a record of the contribution (including all
  personal information I submit with it, including my sign-off) is
  maintained indefinitely and may be redistributed consistent with
  this project or the open source license(s) involved.
-->

#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [x] tests and/or benchmarks are included
- [ ] documentation is changed or added
- [x] commit message and code follows the [Developer's Certification of Origin](https://github.com/fastify/.github/blob/master/CONTRIBUTING.md#developers-certificate-of-origin-11)
      and the [Code of conduct](https://github.com/fastify/.github/blob/master/CODE_OF_CONDUCT.md)