@olivierchatry 2023-12-29
I'm quite sure that BearerType should be case insensitive.