@Uzlopak 2023-12-29
@olivierchatry 

My PR did not change any behaviour regarding case sensitivity. Probably you wanted to comment it on PR #167 by @dancastillo , who changed `bearer` to `Bearer`. 

Eitherway, maybe open an issue to discuss it.