@jsumners 2023-12-29
I am quite certain the spec requires `Bearer` https://www.rfc-editor.org/rfc/rfc6750#section-2.1