@olivierchatry 2023-12-30
https://github.com/lexik/LexikJWTAuthenticationBundle/issues/411

I think this is just a string comparaison, and there will most probably be bearer and Bearer in the wilde. This is really just string comparison, so I'm not sure why not do it.