@Uzlopak 2023-12-30
@olivierchatry 

Please open an issue and discuss this there. Please also refer also to RFCs. This PR did not change that behaviour as I already explained.